pyproject.toml
[project]
name = "langchain-google-firestore"
version = "0.0.1"
description = "LangChain integrations for Google Cloud Firestore"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
dependencies = [
    "langchain==0.1.1",
    "google-cloud-firestore==2.14.0",
    "more_itertools=10.2.0"
]

[project.urls]
Homepage = "https://github.com/googleapis/langchain-google-firestore-python"
Repository = "https://github.com/googleapis/langchain-google-firestore-python.git"
"Bug Tracker" = "https://github.com/googleapis/langchain-google-firestore-python/issues"
Changelog = "https://github.com/googleapis/langchain-google-firestore-python/blob/main/CHANGELOG.md"

[project.optional-dependencies]
test = [
    "black==23.12.0",
    "black[jupyter]==23.12.0",
    "isort==5.13.2",
    "mypy==1.7.1",
    "pytest-asyncio==0.23.0",
    "pytest==7.4.4"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
exclude = [
    "owlbot.py"
]
